requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv
python-jose[cryptography]
python-multipart